ERR: Fix build problems because of recent VTK changes from float to double in the API

+++ Utilities/Xdmf/vtk/vtkXdmfReader.cxx
@@ -446,7 +446,7 @@ void vtkXdmfReader::Execute()
       ( Geometry->GetGeometryType() == XDMF_GEOMETRY_XYZ ) ||
       ( Geometry->GetGeometryType() == XDMF_GEOMETRY_XY ) )
       {
-      float *pp;
+      double *pp;
       XdmfInt64   Length;
       vtkPoints   *Points;
       vtkPointSet *Pointset = ( vtkPointSet *)this->Outputs[idx];
